Split one big test file to three files
Split according to the individual GraphQL mutations that are tested.
