+++ profiles/tests/gdpr/test_gql_delete_my_profile_mutation.py
@@ -1,39 +1,24 @@
-import json
 from string import Template
 
 import pytest
 import requests
-from django.utils.translation import gettext as _
 
 from open_city_profile.consts import (
     CONNECTED_SERVICE_DELETION_NOT_ALLOWED_ERROR,
     MISSING_GDPR_API_TOKEN_ERROR,
     PROFILE_DOES_NOT_EXIST_ERROR,
-    SERVICE_CONNECTION_DOES_NOT_EXIST_ERROR,
     SERVICE_GDPR_API_UNKNOWN_ERROR,
 )
 from open_city_profile.oidc import TunnistamoTokenExchange
 from open_city_profile.tests.asserts import assert_match_error_code
 from profiles.models import Profile
-from profiles.tests.factories import (
-    ProfileFactory,
-    ProfileWithPrimaryEmailFactory,
-    VerifiedPersonalInformationFactory,
-)
+from profiles.tests.factories import ProfileFactory, ProfileWithPrimaryEmailFactory
 from services.models import ServiceConnection
 from services.tests.factories import ServiceConnectionFactory
 from users.models import User
 from utils.keycloak import KeycloakAdminClient
 
 AUTHORIZATION_CODE = "code123"
-
-DOWNLOAD_MY_PROFILE_MUTATION = Template(
-    """
-    {
-        downloadMyProfile(authorizationCode: "${auth_code}")
-    }
-"""
-).substitute(auth_code=AUTHORIZATION_CODE)
 
 DELETE_MY_PROFILE_MUTATION = Template(
     """
@@ -63,52 +48,15 @@ DELETE_MY_PROFILE_MUTATION = Template(
 """
 ).substitute(auth_code=AUTHORIZATION_CODE)
 
-DELETE_MY_SERVICE_DATA_MUTATION = Template(
-    """
-    mutation deleteMyServiceDataMutation($$serviceName: String!, $$dryRun: Boolean) {
-        deleteMyServiceData(
-            input: {
-                authorizationCode: "${auth_code}",
-                serviceName: $$serviceName,
-                dryRun: $$dryRun
-            }
-        ) {
-            result {
-                service {
-                    name
-                    description
-                }
-                dryRun
-                success
-                errors {
-                    code
-                    message {
-                        lang
-                        text
-                    }
-                }
-            }
-        }
-    }
-"""
-).substitute(auth_code=AUTHORIZATION_CODE)
-
 
 def assert_match_error_code_in_results(response, error_code):
     response_data = response["data"]
 
     errors = []
     for name, value in response_data.items():
-        if "results" in value:
-            errors.extend(
-                [
-                    error
-                    for result in value["results"]
-                    for error in result.get("errors", [])
-                ]
-            )
-        if "result" in value:
-            errors.extend(value["result"].get("errors", []))
+        errors.extend(
+            [error for result in value["results"] for error in result.get("errors", [])]
+        )
 
     assert len(errors) > 0
     for error in errors:
@@ -131,221 +79,6 @@ def assert_success_result(response, expected_success=True):
         assert all(success_results)
     else:
         assert not all(success_results)
-
-
-@pytest.mark.parametrize("with_serviceconnection", (True, False))
-def test_user_can_download_profile(
-    user_gql_client, service, mocker, with_serviceconnection
-):
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    service_connection_created_at = None
-    if with_serviceconnection:
-        mocker.patch.object(
-            TunnistamoTokenExchange, "fetch_api_tokens", return_value=None
-        )
-        service_connection = ServiceConnectionFactory(profile=profile, service=service)
-        service_connection_created_at = service_connection.created_at.date().isoformat()
-
-    primary_email = profile.emails.first()
-
-    executed = user_gql_client.execute(DOWNLOAD_MY_PROFILE_MUTATION, service=service)
-
-    if with_serviceconnection:
-        expected_json = json.dumps(
-            {
-                "key": "DATA",
-                "children": [
-                    {
-                        "key": "PROFILE",
-                        "children": [
-                            {"key": "FIRST_NAME", "value": profile.first_name},
-                            {"key": "LAST_NAME", "value": profile.last_name},
-                            {"key": "NICKNAME", "value": profile.nickname},
-                            {"key": "LANGUAGE", "value": profile.language},
-                            {"key": "CONTACT_METHOD", "value": profile.contact_method},
-                            {
-                                "key": "EMAILS",
-                                "children": [
-                                    {
-                                        "key": "EMAIL",
-                                        "children": [
-                                            {
-                                                "key": "PRIMARY",
-                                                "value": primary_email.primary,
-                                            },
-                                            {
-                                                "key": "EMAIL_TYPE",
-                                                "value": primary_email.email_type.name,
-                                            },
-                                            {
-                                                "key": "EMAIL",
-                                                "value": primary_email.email,
-                                            },
-                                        ],
-                                    }
-                                ],
-                            },
-                            {"key": "PHONES", "children": []},
-                            {"key": "ADDRESSES", "children": []},
-                            {
-                                "key": "SERVICE_CONNECTIONS",
-                                "children": [
-                                    {
-                                        "key": "SERVICECONNECTION",
-                                        "children": [
-                                            {"key": "SERVICE", "value": service.name},
-                                            {
-                                                "key": "CREATED_AT",
-                                                "value": service_connection_created_at,
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                        ],
-                    }
-                ],
-            }
-        )
-        assert executed["data"]["downloadMyProfile"] == expected_json, executed
-    else:
-        assert_match_error_code(executed, "PERMISSION_DENIED_ERROR")
-        assert executed["data"]["downloadMyProfile"] is None
-
-
-def download_verified_personal_information_with_loa(
-    loa, user_gql_client, service, mocker
-):
-    profile = VerifiedPersonalInformationFactory(
-        profile__user=user_gql_client.user
-    ).profile
-
-    mocker.patch.object(TunnistamoTokenExchange, "fetch_api_tokens", return_value=None)
-    ServiceConnectionFactory(profile=profile, service=service)
-
-    token_payload = {
-        "loa": loa,
-    }
-    executed = user_gql_client.execute(
-        DOWNLOAD_MY_PROFILE_MUTATION, service=service, auth_token_payload=token_payload
-    )
-
-    full_dump = json.loads(executed["data"]["downloadMyProfile"])
-    profile_dump = next(
-        child for child in full_dump["children"] if child["key"] == "PROFILE"
-    )
-    vpi_dump = next(
-        child
-        for child in profile_dump["children"]
-        if child["key"] == "VERIFIEDPERSONALINFORMATION"
-    )
-
-    return vpi_dump
-
-
-@pytest.mark.parametrize("loa", ["substantial", "high"])
-def test_verified_personal_information_is_included_in_the_downloaded_profile_when_loa_is_high_enough(
-    loa, user_gql_client, service, mocker
-):
-    vpi_dump = download_verified_personal_information_with_loa(
-        loa, user_gql_client, service, mocker
-    )
-
-    assert "error" not in vpi_dump
-    assert len(vpi_dump["children"]) > 0
-
-
-@pytest.mark.parametrize("loa", [None, "foo", "low"])
-def test_verified_personal_information_is_replaced_with_an_error_when_loa_is_not_high_enough(
-    loa, user_gql_client, service, mocker
-):
-    vpi_dump = download_verified_personal_information_with_loa(
-        loa, user_gql_client, service, mocker
-    )
-
-    assert vpi_dump == {
-        "key": "VERIFIEDPERSONALINFORMATION",
-        "error": _("No permission to read verified personal information."),
-    }
-
-
-def test_downloading_non_existent_profile_doesnt_return_errors(user_gql_client):
-    executed = user_gql_client.execute(DOWNLOAD_MY_PROFILE_MUTATION)
-
-    assert executed["data"]["downloadMyProfile"] is None
-    assert "errors" not in executed
-
-
-def test_user_can_download_profile_with_connected_services(
-    user_gql_client, service_1, service_2, gdpr_api_tokens, mocker
-):
-    expected = {"key": "SERVICE-1", "children": [{"key": "CUSTOMERID", "value": "123"}]}
-
-    def mock_download_gdpr_data(self, api_token: str):
-        if self.service.name == service_1.name:
-            return expected
-        else:
-            return {}
-
-    mocker.patch.object(
-        ServiceConnection,
-        "download_gdpr_data",
-        autospec=True,
-        side_effect=mock_download_gdpr_data,
-    )
-    mocker.patch.object(
-        TunnistamoTokenExchange, "fetch_api_tokens", return_value=gdpr_api_tokens
-    )
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    ServiceConnectionFactory(profile=profile, service=service_1)
-    ServiceConnectionFactory(profile=profile, service=service_2)
-
-    executed = user_gql_client.execute(DOWNLOAD_MY_PROFILE_MUTATION)
-
-    response_data = json.loads(executed["data"]["downloadMyProfile"])["children"]
-    assert len(response_data) == 2
-    assert expected in response_data
-
-    # Data does not contain the empty response from service_2
-    assert {} not in response_data
-
-
-def test_user_can_download_profile_using_correct_api_tokens(
-    user_gql_client,
-    service_1,
-    service_2,
-    gdpr_api_tokens,
-    api_token_1,
-    api_token_2,
-    mocker,
-):
-    def mock_download_gdpr_data(self, api_token: str):
-        if (self.service.name == service_1.name and api_token == api_token_1) or (
-            self.service.name == service_2.name and api_token == api_token_2
-        ):
-            return {}
-
-        raise Exception("Wrong token used!")
-
-    profile = ProfileFactory(user=user_gql_client.user)
-    ServiceConnectionFactory(profile=profile, service=service_1)
-    ServiceConnectionFactory(profile=profile, service=service_2)
-    mocked_gdpr_download = mocker.patch.object(
-        ServiceConnection,
-        "download_gdpr_data",
-        autospec=True,
-        side_effect=mock_download_gdpr_data,
-    )
-    mocked_token_exchange = mocker.patch.object(
-        TunnistamoTokenExchange, "fetch_api_tokens", return_value=gdpr_api_tokens
-    )
-
-    executed = user_gql_client.execute(DOWNLOAD_MY_PROFILE_MUTATION)
-
-    mocked_token_exchange.assert_called_once()
-    assert mocked_token_exchange.call_args == ((AUTHORIZATION_CODE,),)
-    assert mocked_gdpr_download.call_count == 2
-    assert executed["data"]["downloadMyProfile"]
 
 
 @pytest.mark.parametrize("with_serviceconnection", (True, False))
@@ -766,30 +499,6 @@ def test_invalid_deletion_errors_from_the_service_are_not_returned(
     )
 
 
-def test_service_doesnt_have_gdpr_query_scope_set(
-    user_gql_client, service_1, gdpr_api_tokens, mocker
-):
-    """Missing query scope should make the query skip the service for a given connected profile."""
-    service_1.gdpr_query_scope = ""
-    service_1.save()
-    response = {
-        "key": "SERVICE",
-        "children": [{"key": "CUSTOMERID", "value": "123"}],
-    }
-    mocker.patch.object(ServiceConnection, "download_gdpr_data", return_value=response)
-    mocker.patch.object(
-        TunnistamoTokenExchange, "fetch_api_tokens", return_value=gdpr_api_tokens
-    )
-    profile = ProfileWithPrimaryEmailFactory(user=user_gql_client.user)
-    ServiceConnectionFactory(profile=profile, service=service_1)
-
-    executed = user_gql_client.execute(DOWNLOAD_MY_PROFILE_MUTATION)
-
-    response_data = json.loads(executed["data"]["downloadMyProfile"])["children"]
-    assert len(response_data) == 1
-    assert response not in response_data
-
-
 def test_service_doesnt_have_gdpr_delete_scope_set(
     user_gql_client, service_1, gdpr_api_tokens, mocker
 ):
@@ -812,149 +521,12 @@ def test_service_doesnt_have_gdpr_delete_scope_set(
     assert_match_error_code(executed, CONNECTED_SERVICE_DELETION_NOT_ALLOWED_ERROR)
 
 
-@pytest.mark.parametrize("query_or_delete", ["query", "delete"])
-def test_api_tokens_missing(user_gql_client, service_1, query_or_delete, mocker):
-    """Missing API token for a service connection that has the query/delete scope set, should be an error."""
+def test_api_tokens_missing(user_gql_client, service_1, mocker):
+    """Missing API token for a service connection that has the delete scope set, should be an error."""
     mocker.patch.object(TunnistamoTokenExchange, "fetch_api_tokens", return_value={})
     profile = ProfileFactory(user=user_gql_client.user)
     ServiceConnectionFactory(profile=profile, service=service_1)
 
-    if query_or_delete == "query":
-        executed = user_gql_client.execute(DOWNLOAD_MY_PROFILE_MUTATION)
-    else:
-        executed = user_gql_client.execute(DELETE_MY_PROFILE_MUTATION)
+    executed = user_gql_client.execute(DELETE_MY_PROFILE_MUTATION)
 
     assert_match_error_code(executed, MISSING_GDPR_API_TOKEN_ERROR)
-
-
-@pytest.mark.parametrize("dry_run", [True, False])
-def test_user_can_delete_data_from_a_service(
-    user_gql_client,
-    service_1,
-    service_2,
-    gdpr_api_tokens,
-    mocker,
-    requests_mock,
-    dry_run,
-):
-    mocker.patch.object(
-        TunnistamoTokenExchange, "fetch_api_tokens", return_value=gdpr_api_tokens
-    )
-    profile = ProfileFactory(user=user_gql_client.user)
-    ServiceConnectionFactory(profile=profile, service=service_1)
-    ServiceConnectionFactory(profile=profile, service=service_2)
-
-    service_1_mocker = requests_mock.delete(
-        service_1.get_gdpr_url_for_profile(profile), status_code=204
-    )
-    service_2_mocker = requests_mock.delete(
-        service_2.get_gdpr_url_for_profile(profile), status_code=204
-    )
-    variables = {
-        "serviceName": service_1.name,
-        "dryRun": dry_run,
-    }
-    executed = user_gql_client.execute(
-        DELETE_MY_SERVICE_DATA_MUTATION, variables=variables
-    )
-
-    assert "errors" not in executed
-
-    if dry_run:
-        assert service_1_mocker.call_count == 1
-        assert service_2_mocker.call_count == 0
-        assert service_1_mocker.request_history[0].qs["dry_run"] == ["true"]
-        assert ServiceConnection.objects.count() == 2
-    else:
-        assert service_1_mocker.call_count == 2
-        assert service_2_mocker.call_count == 0
-        assert service_1_mocker.request_history[0].qs["dry_run"] == ["true"]
-        assert not service_1_mocker.request_history[1].text
-        assert ServiceConnection.objects.count() == 1
-        assert ServiceConnection.objects.first().service == service_2
-
-
-@pytest.mark.parametrize(
-    "errors_from_service",
-    [None, {"errors": [{"code": "ERROR_CODE", "message": {"en": "Error"}}]}],
-)
-def test_error_is_returned_when_service_returns_errors(
-    user_gql_client,
-    service_1,
-    gdpr_api_tokens,
-    mocker,
-    requests_mock,
-    errors_from_service,
-):
-    mocker.patch.object(
-        TunnistamoTokenExchange, "fetch_api_tokens", return_value=gdpr_api_tokens
-    )
-    profile = ProfileFactory(user=user_gql_client.user)
-    ServiceConnectionFactory(profile=profile, service=service_1)
-
-    service_1_mocker = requests_mock.delete(
-        service_1.get_gdpr_url_for_profile(profile),
-        status_code=403,
-        json=errors_from_service,
-    )
-    executed = user_gql_client.execute(
-        DELETE_MY_SERVICE_DATA_MUTATION, variables={"serviceName": service_1.name}
-    )
-
-    if errors_from_service is None:
-        assert_match_error_code_in_results(executed, SERVICE_GDPR_API_UNKNOWN_ERROR)
-    else:
-        assert_match_error_code_in_results(
-            executed, errors_from_service["errors"][0]["code"]
-        )
-
-    assert service_1_mocker.call_count == 1
-    assert ServiceConnection.objects.count() == 1
-    assert ServiceConnection.objects.first().service == service_1
-
-
-def test_error_when_trying_to_delete_data_from_a_service_the_user_is_not_connected_to(
-    user_gql_client, service_1, service_2
-):
-    profile = ProfileFactory(user=user_gql_client.user)
-    ServiceConnectionFactory(profile=profile, service=service_1)
-
-    variables = {
-        "serviceName": service_2.name,
-        "dryRun": False,
-    }
-    executed = user_gql_client.execute(
-        DELETE_MY_SERVICE_DATA_MUTATION, variables=variables
-    )
-
-    assert_match_error_code(executed, SERVICE_CONNECTION_DOES_NOT_EXIST_ERROR)
-
-
-def test_error_when_trying_to_delete_data_from_an_unknown_service(
-    user_gql_client, service_1
-):
-    profile = ProfileFactory(user=user_gql_client.user)
-    ServiceConnectionFactory(profile=profile, service=service_1)
-
-    variables = {
-        "serviceName": "unknown_service",
-        "dryRun": False,
-    }
-    executed = user_gql_client.execute(
-        DELETE_MY_SERVICE_DATA_MUTATION, variables=variables
-    )
-
-    assert_match_error_code(executed, SERVICE_CONNECTION_DOES_NOT_EXIST_ERROR)
-
-
-def test_error_when_using_service_delete_with_non_existent_profile(user_gql_client):
-    variables = {
-        "serviceName": "n/a",
-    }
-    executed = user_gql_client.execute(
-        DELETE_MY_SERVICE_DATA_MUTATION, variables=variables
-    )
-
-    expected_data = {"deleteMyServiceData": None}
-    assert executed["data"] == expected_data
-    assert_match_error_code(executed, PROFILE_DOES_NOT_EXIST_ERROR)
